+++ robotframework-ls/docs/changelog.md
@@ -7,9 +7,10 @@ NEXT
 - New line customization:
     - If a line starts with `#` if a new line is entered before the end of the line a `#` is added in the new line
     - If a line is split a continuation (`...`) is added.
-- Quick fix: create a `local variable` from an `undefined variable`.
-- Quick fix: create a `variable in the variables section` from an `undefined variable`.
-- Extract `local variable` refactoring.
+- Quick fix: create `local variable` from an `undefined variable`.
+- Quick fix: create `variable in the variables section` from an `undefined variable`.
+- Quick fix: create `argument` from an `undefined variable`.
+- Refactoring: extract `local variable`.
 
 
 New in 1.7.3 (2022-12-16)